+++ src/DataAccessLibrary/ISqlDataAccess.cs
@@ -1,5 +1,4 @@
-﻿
-namespace DataAccessLibrary
+﻿namespace DataAccessLibrary
 {
     public interface ISqlDataAccess
     {
