+++ src/DataAccessLibrary/Models/EventModel.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DataAccessLibrary.Models
+﻿namespace DataAccessLibrary.Models
 {
     public class EventModel
     {
