+++ src/DataAccessLibrary/Models/InvitationModel.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DataAccessLibrary.Models
+﻿namespace DataAccessLibrary.Models
 {
     public class InvitationModel
     {
@@ -17,6 +11,5 @@ namespace DataAccessLibrary.Models
         public bool IsActive { get; set; }
         public string Message { get; set; }
         public bool Responded { get; set; }
-
     }
 }
